verl/requirements_sglang.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
ray[default]>=2.10
tensordict<=0.6.2
torchdata
torchvision
transformers
sglang[all]==0.4.6.post5
torch-memory-saver>=0.0.5
